uboot-compile

+++ u-boot/arch/x86/include/asm/mpspec.h
@@ -224,9 +224,9 @@ struct mp_ext_compat_address_space {
  * @mc:		configuration table header address
  * @return:	configuration table end address
  */
-static inline ulong mp_next_mpc_entry(struct mp_config_table *mc)
+static inline u32 mp_next_mpc_entry(struct mp_config_table *mc)
 {
-	return (ulong)mc + mc->mpc_length;
+	return (u32)mc + mc->mpc_length;
 }
 
 /**
@@ -254,9 +254,9 @@ static inline void mp_add_mpc_entry(struct mp_config_table *mc, uint length)
  * @mc:		configuration table header address
  * @return:	configuration table end address
  */
-static inline ulong mp_next_mpe_entry(struct mp_config_table *mc)
+static inline u32 mp_next_mpe_entry(struct mp_config_table *mc)
 {
-	return (ulong)mc + mc->mpc_length + mc->mpe_length;
+	return (u32)mc + mc->mpc_length + mc->mpe_length;
 }
 
 /**
@@ -456,6 +456,6 @@ int mp_determine_pci_dstirq(int bus, int dev, int func, int pirq);
  * @addr:	start address to write MP table
  * @return:	end address of MP table
  */
-ulong write_mp_table(ulong addr);
+u32 write_mp_table(u32 addr);
 
 #endif /* __ASM_MPSPEC_H */
